requirements.txt
alembic==1.7.5
astroid==2.8.4
autopep8==1.6.0
black==21.12b0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
distlib==0.3.4
filelock==3.4.2
flake8==4.0.1
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
graphviz==0.18
greenlet==1.1.2
isort==5.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.5.2
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
pathspec==0.9.0
pipenv==2022.1.8
platformdirs==2.4.0
protobuf==3.19.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.11.1
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
spark-parser==1.8.9
SQLAlchemy==1.4.29
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
uncompyle6==3.8.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
wrapt==1.13.3
xdis==6.0.3
